Make vagrant play nice with OpenStack.
Right now, this is only tested against the stable/folsom branch of OpenStack. Unfortunately, the latest version of fog has trouble authenticating against stable/folsom, so we use fog-1.6.0. That means that a few small things needed to be backported.

We also catch and ignore "Network unreachable" errors after creating a VM, since this error seems to appear consistently and just for a short time.