Added support for nics configurations to allow for selection of a specific network on creation
A new network configuration option has been added that will allow for
an instance to be configured to use a specific network. If a network is
provided when the ssh_info it read the network name will be used to retrieve
the correct ip address. On openstack connect a reference to the network
endpoint has been added if a network configuration is provided. Nics support
was added in fog 1.10.1 so the gemspec has been updated to reflect this
requirment.