Added docs, removed metadata validation and bumped version
I added documentation for the metadata and network additions I made, which I
should have done in the first place. I removed the metadata validation since
it was causing issues when not present. I added some additional output in create
server when fetching networks and suppressed the network and security messages
if no configuration values are provided. I bumped the version to 0.1.4 to take
into account the security/tenet additions made by @jkburges.