Adds support to determine IP address to use
This patch will try to find a valid IP address to use in the event that
a valid network name was not provided in the Vagrantfile. In addition,
it adds a small block of code to print out available network names for
debugging purposes.

This patch tries to retain backwards compatibility with previous
configurations.

This patch should fix issues such as #16