Add proxy option to configuration.
This adds a proxy option to the configuration. In the Vagrantfile we can
override this with a specific value and have it passed down to the Fog
API connection options.

The initialization of the Connection object takes a params hash with
the :proxy attribute:
https://github.com/fog/fog/blob/master/lib/fog/core/connection.rb#L22

This is wired up from both the Compute and Network OpenStack instances:
https://github.com/fog/fog/blob/master/lib/fog/openstack/compute.rb#L311
https://github.com/fog/fog/blob/master/lib/fog/openstack/network.rb#L205